requirements.txt
fastapi
uvicorn
openai
python-dotenv
authlib
sqlalchemy
itsdangerous
python-multipart
slowapi